requirements.txt

tornado
peewee
pymysql
aiomysql
aiofiles
aioredis
loguru
arrow==0.15.0
peewee-async==0.6.0a0
cryptokit
WTForms
wtforms_tornado
django==2.0.7
djangorestframework==3.8.2